requirements.txt
ttkbootstrap>=1.10.0
Pillow>=9.0.0
requests>=2.28.0
pystray>=0.19.0
